requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
psycopg[binary]==3.2.6
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
pydantic-settings==2.8.1
